[refactor] 공통 부모 entity, enumtype, exception, 공통 루트 response -> base 패키지로 이동 - #94
