+++ novelservice/src/main/java/com/iucyh/novelservice/base/entity/SoftDeleteEntity.java
@@ -1,4 +1,4 @@
-package com.iucyh.novelservice.common.entity;
+package com.iucyh.novelservice.base.entity;
 
 import jakarta.persistence.MappedSuperclass;
 import lombok.Getter;
