+++ novelservice/src/main/java/com/iucyh/novelservice/base/enumtype/valuedenum/ValuedEnum.java
@@ -1,0 +1,6 @@
+package com.iucyh.novelservice.base.enumtype.valuedenum;
+
+public interface ValuedEnum {
+
+    String getValue();
+}
