--- novelservice/src/main/java/com/iucyh/novelservice/common/enumtype/valuedenum/ValuedEnum.java
@@ -1,6 +1,0 @@
-package com.iucyh.novelservice.common.enumtype.valuedenum;
-
-public interface ValuedEnum {
-
-    String getValue();
-}
