+++ novelservice/src/main/java/com/iucyh/novelservice/common/exception/DataNotFound.java
@@ -1,10 +1,11 @@
 package com.iucyh.novelservice.common.exception;
 
+import com.iucyh.novelservice.base.exception.BusinessException;
 import com.iucyh.novelservice.common.exception.errorcode.CommonErrorCode;
 
 import java.util.Map;
 
-public class DataNotFound extends ServiceException {
+public class DataNotFound extends BusinessException {
 
     /**
      * @param publicId 해당 리소스를 조회했을 당시 사용한 식별자
