--- novelservice/src/main/java/com/iucyh/novelservice/common/exception/ServiceException.java
@@ -1,31 +1,0 @@
-package com.iucyh.novelservice.common.exception;
-
-import com.iucyh.novelservice.common.exception.errorcode.ErrorCode;
-import lombok.Getter;
-
-import java.util.Map;
-
-@Getter
-public abstract class ServiceException extends RuntimeException {
-
-    private final ErrorCode errorCode;
-    private final Map<String, Object> causes;
-
-    protected ServiceException(ErrorCode errorCode, String message, Map<String, Object> causes) {
-        super(message);
-        this.errorCode = errorCode;
-        this.causes = causes;
-    }
-
-    protected ServiceException(ErrorCode errorCode, String message) {
-        this(errorCode, message, null);
-    }
-
-    protected ServiceException(ErrorCode errorCode, Map<String, Object> causes) {
-        this(errorCode, errorCode.getDefaultMessage(), causes);
-    }
-
-    protected ServiceException(ErrorCode errorCode) {
-        this(errorCode, errorCode.getDefaultMessage(), null);
-    }
-}
