+++ novelservice/src/main/java/com/iucyh/novelservice/episode/web/controller/EpisodeController.java
@@ -1,8 +1,8 @@
 package com.iucyh.novelservice.episode.web.controller;
 
 import com.iucyh.novelservice.common.response.PageWithOffsetResponse;
-import com.iucyh.novelservice.common.response.api.ApiResponseMapper;
-import com.iucyh.novelservice.common.response.api.SuccessResponse;
+import com.iucyh.novelservice.base.response.api.ApiResponseMapper;
+import com.iucyh.novelservice.base.response.api.SuccessResponse;
 import com.iucyh.novelservice.episode.service.EpisodeQueryService;
 import com.iucyh.novelservice.episode.service.EpisodeService;
 import com.iucyh.novelservice.episode.service.dto.command.CreateEpisodeCommand;
