+++ novelservice/src/main/java/com/iucyh/novelservice/novel/exception/DuplicateNovelTitle.java
@@ -1,11 +1,11 @@
 package com.iucyh.novelservice.novel.exception;
 
-import com.iucyh.novelservice.common.exception.ServiceException;
+import com.iucyh.novelservice.base.exception.BusinessException;
 import com.iucyh.novelservice.novel.exception.errorcode.NovelErrorCode;
 
 import java.util.Map;
 
-public class DuplicateNovelTitle extends ServiceException {
+public class DuplicateNovelTitle extends BusinessException {
 
     public DuplicateNovelTitle(String title) {
         super(
