+++ novelservice/src/main/java/com/iucyh/novelservice/novel/exception/InvalidNovelCursor.java
@@ -1,9 +1,9 @@
 package com.iucyh.novelservice.novel.exception;
 
-import com.iucyh.novelservice.common.exception.ServiceException;
+import com.iucyh.novelservice.base.exception.BusinessException;
 import com.iucyh.novelservice.novel.exception.errorcode.NovelErrorCode;
 
-public class InvalidNovelCursor extends ServiceException {
+public class InvalidNovelCursor extends BusinessException {
 
     public InvalidNovelCursor() {
         super(NovelErrorCode.INVALID_CURSOR);
