+++ novelservice/src/main/java/com/iucyh/novelservice/novel/exception/NovelAlreadyCompleted.java
@@ -1,9 +1,9 @@
 package com.iucyh.novelservice.novel.exception;
 
-import com.iucyh.novelservice.common.exception.ServiceException;
+import com.iucyh.novelservice.base.exception.BusinessException;
 import com.iucyh.novelservice.novel.exception.errorcode.NovelErrorCode;
 
-public class NovelAlreadyCompleted extends ServiceException {
+public class NovelAlreadyCompleted extends BusinessException {
 
     public NovelAlreadyCompleted() {
         super(NovelErrorCode.NOVEL_ALREADY_COMPLETED);
