+++ novelservice/src/main/java/com/iucyh/novelservice/novel/exception/NovelAlreadyHasPrologue.java
@@ -1,11 +1,9 @@
 package com.iucyh.novelservice.novel.exception;
 
-import com.iucyh.novelservice.common.exception.ServiceException;
+import com.iucyh.novelservice.base.exception.BusinessException;
 import com.iucyh.novelservice.novel.exception.errorcode.NovelErrorCode;
 
-import java.util.Map;
-
-public class NovelAlreadyHasPrologue extends ServiceException {
+public class NovelAlreadyHasPrologue extends BusinessException {
 
     public NovelAlreadyHasPrologue() {
         super(NovelErrorCode.NOVEL_ALREADY_HAS_PROLOGUE);
