+++ novelservice/src/main/java/com/iucyh/novelservice/novel/exception/NovelHasNoCommonEpisodes.java
@@ -1,9 +1,9 @@
 package com.iucyh.novelservice.novel.exception;
 
-import com.iucyh.novelservice.common.exception.ServiceException;
+import com.iucyh.novelservice.base.exception.BusinessException;
 import com.iucyh.novelservice.novel.exception.errorcode.NovelErrorCode;
 
-public class NovelHasNoCommonEpisodes extends ServiceException {
+public class NovelHasNoCommonEpisodes extends BusinessException {
 
     public NovelHasNoCommonEpisodes() {
         super(NovelErrorCode.NOVEL_HAS_NO_COMMON_EPISODES);
